backend/requirements.txt
aiofiles==0.4.0
certifi==2019.9.11
chardet==3.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
multidict==4.5.2
PyJWT==1.7.1
pymongo==3.9.0
python-dotenv==0.10.3
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
Sanic-Cors==0.9.9.post3
Sanic-Plugins-Framework==0.8.2
ujson==1.35
urllib3==1.25.6
uvloop==0.13.0
websockets==8.0.2
